feat: cleanup different workflows for rules fetching
