--- src/app/(layout-with-navigation)/_components/Navigation.tsx
@@ -1,114 +1,0 @@
-'use client'
-
-// import { loadPreviousSimulation, resetLocalisation } from '@/actions/actions'
-import ProgressCircle from '@/design-system/utils/ProgressCircle'
-import CardGameIcon from '../../../components/icons/CardGameIcon'
-
-import Logo from '@/components/misc/Logo'
-import Trans from '@/components/translation/Trans'
-
-import GroupIcon from '@/components/icons/GroupIcon'
-import PersonIcon from '@/components/icons/PersonIcon'
-import { useDebug } from '@/hooks/useDebug'
-import { useUser } from '@/publicodes-state'
-import { capitalizeString } from '@/utils/capitalizeString'
-import ModelVersionIndicator from './navigation/ModelVersionIndicator'
-import NavLink from './navigation/NavLink'
-import PRIndicator from './navigation/PRIndicator'
-
-const ActionsInteractiveIcon = ({ className = '' }) => {
-  const actionChoices = {}
-  const count = Object.values(actionChoices).filter((a) => a === true).length
-  return <CardGameIcon className={className} number={count} />
-}
-
-const openmojis = {
-  test: '25B6',
-  action: 'E10C',
-  conference: '1F3DF',
-  sondage: '1F4CA',
-  profile: 'silhouette',
-  silhouettes: 'silhouettes',
-  personas: '1F465',
-  github: 'E045',
-}
-
-export const openmojiURL = (name: keyof typeof openmojis) =>
-  `/images/misc/${openmojis[name]}.svg`
-
-export const actionImg = openmojiURL('action')
-export const conferenceImg = openmojiURL('conference')
-
-export default function Navigation() {
-  const isDebug = useDebug()
-  const { getCurrentSimulation } = useUser()
-  const enquete = ''
-  const persona: string | undefined = getCurrentSimulation()?.persona
-
-  return (
-    <nav
-      id="mainNavigation"
-      className="z-50 my-2 flex h-auto flex-col justify-center pb-8 outline-none lg:sticky lg:top-0 lg:my-4 lg:w-[14rem] lg:shrink-0 lg:justify-start lg:overflow-hidden lg:border-0 lg:border-r-[1px] lg:border-solid lg:border-grey-200">
-      <Logo size="sm" className="hidden lg:block" />
-      <div className="z-100 fixed bottom-0 left-0 m-0 w-screen border-0 border-t-[1px] border-solid border-grey-200 lg:static lg:z-auto lg:mt-4 lg:w-auto lg:border-none">
-        <ul className="m-0 flex h-20 w-full list-none justify-between bg-white py-1 shadow-md sm:px-4 lg:h-auto lg:flex-col lg:justify-start lg:gap-1 lg:bg-none lg:py-2 lg:shadow-none">
-          <NavLink
-            href="/simulateur/bilan"
-            className="justify-end !p-0 text-default lg:justify-start lg:!p-4">
-            <ProgressCircle className="lg:mr-4" />
-            <span className="text-base md:text-lg">
-              <Trans>Le test</Trans>
-            </span>
-          </NavLink>
-
-          <NavLink
-            href="/actions"
-            className="justify-end !p-0 text-default lg:justify-start lg:!p-4">
-            <ActionsInteractiveIcon className="w-12 lg:mr-4" />
-
-            <span className="text-base md:text-lg">
-              <Trans>Agir</Trans>
-            </span>
-          </NavLink>
-
-          {!enquete && (
-            <NavLink href="/profil" className="text-default">
-              <div className="relative">
-                <PersonIcon className="mr-[20px]" />
-              </div>
-              <span className="text-base  md:text-lg">
-                {!persona ? (
-                  <Trans>Profil</Trans>
-                ) : (
-                  <span className="rounded-md bg-primary-500 px-4 py-2 text-white">
-                    {capitalizeString(persona.split(' . ')[1])}
-                  </span>
-                )}
-              </span>
-            </NavLink>
-          )}
-
-          {!enquete && (
-            <NavLink
-              href="/amis"
-              data-cypress-id="amis-link"
-              className="text-default">
-              <GroupIcon className="mr-[20px]" />
-
-              <span className="text-base md:text-lg">
-                <Trans>Amis</Trans>
-              </span>
-            </NavLink>
-          )}
-          <PRIndicator />
-          <ModelVersionIndicator />
-          {isDebug ? (
-            <div className="mx-auto hidden rounded-lg bg-red-600 px-4 py-2 text-center font-bold uppercase text-white lg:block">
-              Debug
-            </div>
-          ) : null}
-        </ul>
-      </div>
-    </nav>
-  )
-}
