+++ src/hooks/useDataServer.ts
@@ -1,8 +1,4 @@
-import {
-  DATA_SERVER_URL,
-  NGC_MODEL_API_URL,
-  getModelPRUrl,
-} from '@/constants/urls'
+import { NGC_MODEL_API_URL, getModelPRUrl } from '@/constants/urls'
 import useModelVersion from '@/hooks/useModelVersion'
 import { usePRNumber } from './usePRNumber'
 
@@ -17,7 +13,4 @@ export const useDataServer = () => {
   }
 
   return `${NGC_MODEL_API_URL}/${modelVersion}`
-  if (process.env.NEXT_PUBLIC_LOCAL_DATA) return
-
-  return DATA_SERVER_URL
 }
